Run extract class refactoring on ImageList
